+++ CHANGELOG.md
@@ -7,6 +7,7 @@ this changelog also includes the original WARDEN changelog.
 ## 1.0.0-SNAPSHOT
 * More proper exception handling
 * More lenient hex parsing: also filter out `:` in addition to whitespaces
+* Add direct link to file bug for logical error
 
 ## 0.9.999
 * Quality of life improvements: